+++ recorder/src/service/qmmf_camera_rescaler.cc
@@ -538,7 +538,8 @@ status_t CameraRescalerMemPool::PopulateMetaInfo(CameraBufferMetaData &info,
       IAllocDevice::AllocDeviceAction::GetAlignedHeight,
       static_cast<void*>(&alignedH));
   if (MemAllocError::kAllocOk != ret) {
-    QMMF_ERROR("%s: Unable to query stride&scanline: %d\n", __func__, ret);
+    QMMF_ERROR("%s: Unable to query stride&scanline: %d\n", __func__,
+      (int32_t) ret);
     return BAD_VALUE;
   }
 
@@ -546,7 +547,8 @@ status_t CameraRescalerMemPool::PopulateMetaInfo(CameraBufferMetaData &info,
       IAllocDevice::AllocDeviceAction::GetAlignedWidth,
       static_cast<void*>(&alignedW));
   if (MemAllocError::kAllocOk != ret) {
-    QMMF_ERROR("%s: Unable to query stride&scanline: %d\n", __func__, ret);
+    QMMF_ERROR("%s: Unable to query stride&scanline: %d\n", __func__,
+      (int32_t) ret);
     return BAD_VALUE;
   }
